Remove Apache Derby
