--- src/main/scala/dk/aau/modelardb/engines/derby/DerbyStorage.scala
@@ -1,23 +1,0 @@
-/* Copyright 2018-2021 Aalborg University
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package dk.aau.modelardb.engines.derby
-
-import org.apache.derby.vti.Restriction
-import dk.aau.modelardb.core.{SegmentGroup, Storage}
-
-trait DerbyStorage extends Storage {
-  def storeSegmentGroups(segments: Array[SegmentGroup], length: Int): Unit
-  def getSegmentGroups(filter: Restriction): Iterator[SegmentGroup]
-}
